Main code to complete #45, #49, #44, #43, #42
THIS IS THE PROJECT LIBRARY FOR ORIGAMI-ELECTRONICS!!!!

Base implementation which calls all library functions with ONLY a print-statement output is functional.
So far, without being able to fully operate the different components in tandem, the code is not fully operational. If you focus on 1 component and comment out the code involving other components, the focused component's code will run. 

**Note:** 
1. I already obtain code that properly operates the Buzzer as I intend. This code was not made public since it was created before the creation of the Repository. While working on the project, I have been attempting to improve/expand the previously coded file for the Buzzer. Unfortunately, I ended breaking my code and need to further improve my implementations.
2. Besides the buzzer code not running correctly, the other components's code can be run correctly.